[REFACTOR]: 키보드 내리기를 앱 전역에서 화면별 적용으로 전환
dismissKeyboardOnTap을 DesignSystem 공용 extension으로 옮기고
KakaoMap POI 탭 제스처 충돌을 피하기 위해 TextField 화면에만 개별 적용
