+++ Projects/App/Sources/Application/BangawoApp.swift
@@ -37,7 +37,6 @@ struct BangawoApp: App {
     var body: some Scene {
         WindowGroup {
             RootView(store: store)
-                .dismissKeyboardOnTap()
                 .onOpenURL { url in
                     if AuthApi.isKakaoTalkLoginUrl(url) {
                         AuthController.handleOpenUrl(url: url)
@@ -69,19 +68,4 @@ struct BangawoApp: App {
                 )
                 Log.debug("👤 [Naver] Login SDK 초기화 완료")
         }
-}
-
-private extension View {
-    func dismissKeyboardOnTap() -> some View {
-        simultaneousGesture(
-            TapGesture().onEnded {
-                UIApplication.shared.sendAction(
-                    #selector(UIResponder.resignFirstResponder),
-                    to: nil,
-                    from: nil,
-                    for: nil
-                )
-            }
-        )
-    }
 }
