+++ Projects/Presentation/AuthFlowFeature/Sources/DepartureSearch/DepartureSearchView.swift
@@ -57,6 +57,7 @@ public struct DepartureSearchView: View {
             isToastVisible = false
             Task { @MainActor in isToastVisible = true }
         }
+        .dismissKeyboardOnTap()
     }
 }
